+++ app/dash/[path]/layout.tsx
@@ -1,4 +1,4 @@
-import "../globals.css";
+import "../../globals.css";
 import { config } from "@fortawesome/fontawesome-svg-core";
 import "@fortawesome/fontawesome-svg-core/styles.css";
 config.autoAddCss = false;
